IM.Base - Se agrega la clase Invitations para el manejo de invitaciones - Se agrega el metodo IsInRangeDate en la clase DateHelper - Se cambian los parametros a decimales en el metodo ValidateNumber() de la clase ValidateHelper
IM.BusinessRules
- Se crea el BRShows con el siguiente metodo
    GetValidateShow()
- Se crea el BRShowSalesman con los siguientes metodos
    GetShowsSalesmen()
    SaveShowsSalesmen()

IM.Host
- se Crea la ventana  frmShowsSalesmen
- Se agregan mas catalogos al metodo GetAllCatalogsHost en frmHost
- se suben avances de la ventana Show

+++ IM.Base/Helpers/ValidateHelper.cs
@@ -1,11 +1,9 @@
 ﻿using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Media;
 using Xceed.Wpf.Toolkit;
 
 namespace IM.Base.Helpers
@@ -388,8 +386,11 @@ namespace IM.Base.Helpers
     /// <param name="upperBound">Limite superior</param>
     /// <param name="description">Nombre del campo</param>
     /// <returns>True is Valid | False No</returns>
+    /// <history>
     /// [erosado] 06/08/2016  Created.
-    public static bool ValidateNumber(int number, int lowerBound, int upperBound, string description)
+    /// [aalcocer] 11/08/2016 Modified. Se cambian los parametros a decimales
+    /// </history>
+    public static bool ValidateNumber(decimal number, decimal lowerBound, decimal upperBound, string description)
     {
       bool _isValid = true;
       if (!(lowerBound <= number && number <= upperBound))
